[DO NOT MERGE] Fix the way FreeType headers are incldued.
FreeType headers have to be included with macros defined in ftheaders.h.
Not doing so breaks when the FreeType header location changes as was the case
between FreeType 2.6.0 and 2.6.2.

This issue blocks updating FreeType to 2.6.2.

Fixed in the upstream:
 https://pdfium.googlesource.com/pdfium/+/56461d66054c53fe3abc88208eaf5d84e80e6551

This CL is back-porting the following AOSP master CL:
  https://android-review.googlesource.com/#/c/196608/

BUG: 24296662

Change-Id: I174a6ad2d3efcce5a68317957869ae898869f850
Signed-off-by: Jungshik Shin <jungshik@google.com>
(cherry picked from commit 5022fbe8a4ab6f58278e7ccefd765f339488aed5)